+++ NEWS.md
@@ -1,3 +1,8 @@
+# quanteda.sentiment 0.2x
+
+* Usability improvements
+* nested_scope = "dictionary" is now standard for applications on tokens, for textstat_polarity() (#12)
+
 # quanteda.sentiment 0.1
 
 * Added a `NEWS.md` file to track changes to the package.